Create #ternari.c
TERNARY OPERATOR [ ? : ] 

It is also called as Conational Operator. This statement is used in case when any if and else part have only one statement but must not more than one. We use colon operator in between two statements.

SYNTAX OF if STATEMENT.

if( condition )
       Statement - 1;
else
      Statement - 2;

We can do the same using Ternary / Conditional operator as follow.

SYNTAX.

Condition ? Statement -1 : Statement – 2 ;
Here, Condition is evaluated first. If it is evaluate to TRUE (1) then Statement – 1 is evaluated and this becomes
value of Conditional Expression. If it is evaluate to FALSE (0) then Statement – 2 is evaluated and this becomes
value of Conditional Expression.

EXAMPLE.

No1 > No2 ? MAX = No1 : MAX = No2 ;