Create #factorial.c
The factorial of any natural number “n” is defined as the product of all natural numbers till n. It plays an important role in various mathematical concepts such as permutations, combinations, probability, and many others.

The notation of the factorial function is “!” or “⌋”. If we have to find the factorial of the number n then, it is written as n! or n⌋. Let’s understand it with some examples:

0! = 1 ( Value of Factorial 0 is 1 because it shows the number of possible ways to arrange a data set with no value in it is 1)

1! = 1
2! = 2 × 1 = 2
3! = 3 × 2 × 1 = 6
4! = 4 × 3 × 2 × 1 = 24
5! = 5 × 4 × 3 × 2 × 1 = 120